+++ src/main/java/com/sdl/selenium/utils/config/WebLocatorConfigReader.java
@@ -10,7 +10,8 @@ public class WebLocatorConfigReader extends PropertiesReader {
     private static final String DEFAULT_CONFIG = ""+
             "\n weblocator.log.containers=true" +
             "\n weblocator.log.useClassName=false" +
-            "\n weblocator.log.logXPathEnabled=true" +
+            "\n weblocator.log.xPath=true" +
+            "\n weblocator.log.suggestions=true" +
             "\n weblocator.highlight=false" +
             "\n weblocator.defaults.renderMillis=3000" +
             "\n #accepted values for searchType: " + Arrays.asList(SearchType.values()) +